Finalized. Tested to work.

+++ include/embedder.h
@@ -5,7 +5,7 @@
 
 class EmbeddingClient {
 public:
-  EmbeddingClient(const std::string &url, int timeout = 30000);
+  EmbeddingClient(const std::string &url, size_t timeout = 30000);
   void generateEmbeddings(const std::vector<std::string> &texts, std::vector<float> &embedding);
   std::vector<std::vector<float>> generateBatchEmbeddings(const std::vector<std::string> &texts);
 
@@ -14,7 +14,7 @@ private:
   std::string host_;
   std::string path_;
   int port_;
-  int timeoutMs_;
+  size_t timeoutMs_;
 
   void parseUrl();
   float calculateL2Norm(const std::vector<float> &vec);
